// src/exceptions/types/logic-exceptions.enum.ts
export enum LogicExceptionList {
  USER_NOT_FOUND = 'user_not_found',
  ACCESS_DENIED = 'access_denied',
  ITEM_ALREADY_EXIST = 'item_already_exist',
  ITEM_NOT_FOUND = 'item_not_found',
  WORK_TYPE_ALREADY_EXIST = 'work_type_already_exist',
  WORK_TYPE_NOT_FOUND = 'work_type_not_found',
  REVIEW_NOT_FOUND = 'review_not_found',
  EMPLOYEE_ITEM_NOT_FOUND = 'employee_item_not_found',
  UNAUTHORIZED = 'unauthorized',
  FILE_TOO_LARGE = 'file_too_large',
  FILE_UNSUPPORTED_TYPE = 'file_unsupported_type',
  ORDER_NOT_FOUND = 'order_not_found',
  ORDER_STATUS_INCORRECT = 'order_status_incorrect',
  EMPLOYEE_ALREADY_ATTACH = 'employee_already_attach',
  REPLY_NOT_FOUND = 'reply_not_found',
  REPLY_ALREADY_EXIST = 'reply_already_exist',
  CHAT_CLOSED = 'chat_closed',
  CHAT_CANNOT_START = 'chat_cannot_start',
  CHAT_ALREADY_EXISTS = 'chat_already_exists',
  CHAT_NOT_FOUND = 'chat_not_found',
  BALANCE_NOT_FOUND = 'balance_not_found',
  PRICE_NOT_FOUND = 'price_not_found',
  INSUFFICIENT_FUNDS = 'insufficient_funds',
  USER_ALREADY_IN_BLACK_LIST = 'user_already_in_black_list',
  BLACK_LIST_NOT_FOUND = 'black_list_not_found',
  RATE_LIMIT_DENIED = 'rate_limit_denied',
  COMMENT_ALREADY_EXISTS = 'comment_already_exists',
}